Split `reporting` into multiple files
